Extract Supervision Tree controls to a LiveComponent
